feat: #178 기존 중복코드 삭제 및 정위치
