+++ Spoony-iOS/Spoony-iOS/Source/Feature/Home/Components/PlaceHolderCardView/RoundedCorner.swift
@@ -1,0 +1,20 @@
+//
+//  RoundedCorner.swift
+//  Spoony-iOS
+//
+//  Created by 이지훈 on 2/10/25.
+//
+
+import SwiftUI
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect,
+                               byRoundingCorners: corners,
+                               cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
